Major: 0.0.101
New Features:
1. Huge refactor:
Now Nerve will use addRawFile -> updateFileInfo structure instead of
the old appendMusicFile function to finished the mission.
This means: we won’t use signals and slots to append files so that
Nerve can be faster, and it is possible that the file’s data can be
updated when user double click it.
Here’re some tips:
* If you want to break a long function call, try signals and slots,
even it will slow your app, it pretty work.
* In multi-thread apps, your const call might caused crashed.
(KNMusicInfoCollector)